requirements.txt
# ============================================================
# requirements.txt — XAUUSD AI Trading Bot
# ============================================================
# Install with: pip install -r requirements.txt
#
# Note: MetaTrader5 is Windows-only.
# All other packages work on Windows/Mac/Linux.

# ── Core data & analysis ─────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── MT5 (Windows only) ───────────────────────────────────────
MetaTrader5>=5.0.45

# ── Machine learning ─────────────────────────────────────────
xgboost>=2.0.0
scikit-learn>=1.3.0

# ── Groq ─────────────────────────────────────────────────────
groq>=0.9.0

# ── Telegram bot ─────────────────────────────────────────────
python-telegram-bot>=20.0

# ── News fetching ────────────────────────────────────────────
requests>=2.31.0
feedparser>=6.0.10

# ── Web scraping (ForexFactory calendar) ─────────────────────
beautifulsoup4>=4.12.0

pytz>=2024.1
